acquire staging lock before sealed lock
It is very important that the staged lock is acquired before the
sealed lock. Both sealing and piece-adding acquire the staged lock
before the sealed lock. If this thread holds the sealing lock but not
ed lock while some other thread holds the staged lock but not
the sealing lock, we deadlock.