fr32: move the pos transformation functions inside `PaddingMap`
Move `transform_bit_pos` and `transform_byte_pos` inside `PaddingMap`, they rely
on the padding semantics, any other transformation call is meaningless outside
that scope. Remove the from/to arguments using instead the `PaddingMap`'s
internal values and set just a flag to indicate in which direction
(padding/unpadding) are we transforming to.