fr32: keep the `missing_data_bits <= source_bits` check
With the sub-byte padding scheme clarified remove the associated `TODO` and
explicitly ask for at least the number of missing data bits in the `source`.
Even though the `<` case is enough (because the extra bits will be converted to
padding) keep the explicit padding at the end of the last chunk for clarity
(leave a `TODO` as a later optimization opportunity if worth it).