fr32: check `len` passed to `write_unpadded`
Do not allow calls to `write_unpadded` with a `len` argument bigger than the
length of raw data actually encoded in what's left from the `source` (starting
at `offset`).

Remove a test that tried to force that situation (which was actually relying on
raw data streams of length that would result in a group of whole elements, i.e.,
no extra bits). Replace it with an explicit test for an error in that case.