add max_sealed_bytes_per_sector
VDF PoSt setup and public parameters need to include SECTOR_SIZE, which is the
number of bytes in the sealed sector. Today, this is just a u64 which we plumb
around. To reduce confusion and the potential for bugs, we add a new method to
SectorConfig and pass the SectorConfig around when generating setup and public
parameters.