refactor(storage_proofs): make fixed node size in drgraph more explicit
* refactor(storage_proofs): make fixed node size in drgraph more explicit

The node size in the depth robust graph was already hard-coded to 32 bytes,
but this value was still used on a few occasions as a magic number. Change
the code to rely on those 32 bytes which is defined by the NODE_SIZE constant.

* refactor(storage_proofs): use constant instead of magic number

The sizes are all about the size of the nodes, hence use the NODE_SIZE constant.