feat(locking): make sys lock handle leaks
The previous locking script didn't clean up leaks. This new iteration places a pid file in the lock directory and, when first run, checks to ensure that any existing lock corresponds to a running process (if not, the script relinquishes the lock on behalf of the now-dead process which acquired it).