fix: prefer u64 instead of usize
In the current code base `usize` is used a lot, but most of the time it is
use as if it were a u64. Hence use u64 for new code.

In the future the code should be checked if u64 sizes are always needed, or
if sometimes smaller sizes would be sufficient.