Run PoSt proofs with high priority
Running proofs with high priority means that they will always run on the GPU.
If the GPU is already used, the lower priority proof will be moved to the CPU.

Do not always run PoSt proofs with high priority, but make it configurable.

Proofs can be forced to run on the CPU instead of the GPU. This is used to
run some proofs with high priority.

The `gpu-cpu-test` tool tests if this actually works. It spawns two threads/
processed run proofs. Those get killed after 5 minutes of running. The overall
test runs longer as some input data needs to be generated. By default, one
thread/process will always be prioritized to run on the GPU. The other one
might be moved to the CPU.

When running:

    $ cd fil-proofs-tooling
    $ RUST_LOG=debug cargo run --release --bin gpu-cpu-test

The return values of the tests might look like:

    2020-02-06T15:35:41.157 22046 main INFO gpu_cpu_test > Thread high info: RunInfo { elapsed: 303.184163413s, iterations: 21 }
    2020-02-06T15:35:45.065 22046 main INFO gpu_cpu_test > Thread low info: RunInfo { elapsed: 307.092464809s, iterations: 7 }

Clearly the high priority thread got more work done.

When running without the "GPU stealing" feature, where one tread demands to
run on the GPU:

    $ RUST_LOG=debug cargo run --release --bin gpu-cpu-test -- --gpu-stealing false

The return values indicate that both threads got the same amount of time on the GPU:

    2020-02-06T16:34:24.015 1252 main INFO gpu_cpu_test > Thread high info: RunInfo { elapsed: 310.90826092s, iterations: 24 }
    2020-02-06T16:34:24.015 1252 main INFO gpu_cpu_test > Thread low info: RunInfo { elapsed: 305.103257599s, iterations: 24 }

The same tests can also be run as separate processes with similar results:

    $ RUST_LOG=debug cargo run --release --bin gpu-cpu-test -- --mode processes
    2020-02-06T17:00:58.360 13488 main INFO gpu_cpu_test > Process high info: Thread high info: RunInfo { elapsed: 300.641724496s, iterations: 24 }
    2020-02-06T17:00:58.360 13488 main INFO gpu_cpu_test > Process low info: Thread high info: RunInfo { elapsed: 310.933157735s, iterations: 9 }

    RUST_LOG=debug cargo run --release --bin gpu-cpu-test -- --mode processes --gpu-stealing false
    2020-02-06T16:47:11.853 7404 main INFO gpu_cpu_test > Process low info: Thread high info: RunInfo { elapsed: 300.857434143s, iterations: 21 }
    2020-02-06T16:48:53.401 7404 main INFO gpu_cpu_test > Process high info: Thread high info: RunInfo { elapsed: 306.301438793s, iterations: 26 }