feat: add support for running multiple low priority threads
Instead of having only a flag to run a single or two threads, change
the `parallel` option to a number, to define how many threads should
be run. The first thread/process is always a high priority one, the
other ones are low priority.

`1` means running only a single high priority proof (this is what
`--parallel false` was), `2` means running a high and a low priority
proof (this is what `--parallel true` was), higher numbers mean
one high priority proof and n-1 low priority ones.