fix(filecoin-proofs): explicitly use 16 byte hash length
Currently we are hashing filenames using blake2 configured with a hash
length of 64 bytes. We are the converting this to hex digits and
truncating to take the first 32 digits - this gives us a string that
represents 16 bytes from the hash.

Blake2 can be configured with a hash length, if we want 16
bytes (in order to maintain the current string length) we should
explicitly configure blake2 to use a 16 byte hash length.

Closes: #869