refactor(filecoin-proofs): remove clone on double reference
Clippy emits error:

        error: using `clone` on a double-reference; this will copy the
        reference instead of cloning the inner type

This error is caused by code that is mildly convoluted by the fact
that the only method on  `TempDir` that returns a `PathBuf` consumes
self. We cannot use that because at the call site we cannot consume
the temp dir. To get around this we do a little dance of getting a
reference to a `Path` and then calling `to_path_buf()` on it. We end
up with code that looks kind of funky:

        self.cache_dire.path().to_path_buf()

That's the best I could come up with :)