feat(storage-proofs): add safe write functionality
Currently if the filesystem fills up during a write we panic. This is
sub-optimal.  It would be nice if we could first allocate space before
writing then only proceed to write if there as enough available space
in the filesystem.

Add a trait `SafeFileExt` that provides two methods

- `safe_write_all`
- `safe_set_len`

These are 'safe' versions of there obvious namesakes.

While we are at it add a function to safely wrap `io::copy` that opens
a file and uses the newly defined `safe_set_len` before doing the copy.