refactor: apply an 'in name only' version bump
This commit is here because v4.0.1 was a hotfix placed directly on the v4.0.0 branch.  Because that branch contains fixes that are not desired in the master branch (or the next version), we are temporarily updating the cargo versions manually to 4.0.1 here just in order to publish the next release.  Mea culpa, this kind of workaround will hopefully not be needed in the future.