refactor: simplify the code that reports faulty sectors
The code was using a parallel iterator, whose result was then processed
sequentially in a for loop. I'm not sure how much parallization that would
give us. Hence I decided to just use a sequntial (non-parallel) iterator
instead.

The alternative would be using a a parallel iterator with mutex around
the vectors, which seems overkill for this use case.