chore: most of CI now runs on the rust-toolchain defined version (#1587)
We use the `rust-toolchain` file in order to run things on a specific
version of Rust. Change the CI to use that version. Prior to this
commit it was using a nightly build, which was needed in the past,
but no longer.

The only exception is on ARM, where we still need nightly.