fix: make builds on stable and aarch64 possible
When using Rust nightly we are able to enable an optimization for aarch64.
With this change it's now possible to compile for aarch64 on stable rust,
while the optimization is still enabled when compiled with nightly.

A similar fix was made for bellperson at
https://github.com/filecoin-project/bellperson/pull/269

This fix can be tested even on x86_64 on Linux with cross compiling
for aarch64:

    cargo build --target aarch64-unknown-linux-gnu

Fixes #1652.