refactor: make encode_into more consistent
The `encode_into()` function takes a `PoRepConfig` and immediately converts it
into a `SectorUpdateConfig`. It's a better and more consistent with existing
APIs (e.g. the `decode_from()`) to take a `SectorUpdateConfig` directly.

BREAKING CHANGE: `encode_into()` not takes a `SectorUpdateConfig` instead of
a `PoRepConfig`. In order to upgrade you can call

    let config = SectorUpdateConfig::from_porep_config(&porep_config);

before calling `encode_into()`.