refactor: clarify `h` and `h_select`
The circuit supports a fixed hard-coded list of `h` values (7-12). Those
values are addressed by their index in the list. It can be seen as a 6-bit
bit-pattern, which is called `h_select`.

The vanilla proof also supports that list of `h` values.

The proofs APIs use a fixed value, 1 for sector sizes <= 32KiB and 10 for
bigger sector sizes.

All this should make the code more consistent and easier to follow.

BREAKING CHANGE: `hs` and `HSelect` are not longer part of the
`filecoin-proofs` public API.