refactor: refactor p_aux/t_aux handling into the util module
`p_aux` and `t_aux` is dealt with throughout the code. In the update proofs API
they are nicely put into functions. Move those functions into the common utility
module and then use those in all places.

The update proofs API puts a restriction on using the Poseidon hasher, which was
also part of those functions. This is no longer the case (although it's correct),
as it would mean breakage of the public API.