fix: re-export more types
Re-export `AggregateVersion`:

`bellperson`s `AggregateVersion` is part of the public API due to the
`CompoundProof` trait. With re-exporting it, it's possible that downstream
dependants like filecoin proofs API don't need to depend on `bellperson`
directly, but can use this re-export instead.

Re-export `filecoin_hashers::Hasher` trait:

The `filecoin_hashers::Hasher` trait is needed a lot in downstream dependants
when the domain of the Tree hasher needs to be accessed. Hence re-export it,
so that those dependents don't need to depend directly on `filecoin_hashers`
only due to this single trait.