refactor: split challenges and layer information (#1730)
At the moment the challenges and the information about the number of layers
is put into a single `LayerChallenges` struct. This misleads into thinking
that the challenge generation is related to the number of layers, but it is
*not*. `LayerChallenges` now only contains the information about challenges.

This makes the code eaier to follow, as it's now clearer, whether a function
needs the layers or the challenges.