fix: remove wrong assert in Synth PoRep code path
Due to me understanding the code wrongly during the refactoring at
https://github.com/filecoin-project/rust-fil-proofs/pull/1720, there
was a wrong assert inserted into the code. With removing this assert
the lifecycle 32GiB test as well as the benchy synthetic proofs
generation works as expected again.

This commit also updates the README with the correct command to run
benchy for synthetic proofs generation.