refactor: move the check that public inputs match stored data
Move the check that that the public inputs match the stored data out of
the `prove_layers_generate()` where it is called on every partition
(with exactly the same data) one level up into `prove_layers()`. This
way the same check is performed only once.

This also makes it possible to pass in the `replica_id` directly into
the `prove_layers_generate()` function. This way we don't need to pass
in the full public inputs, which makes it more obvious which parameters
are passed into the actual proving. It e.g. makes clear that the seed
is only used for the challenge generation, but not for the proving
afterwards (the challenges are of course also part of the proving,
though).