fix: use a fixed seed for Non-interactive PoRep
The Interactive PoRep requires a seed for the proof. The Non-interactive PoRep
does not. In order to keep the public APIs simple, just make sure that the
passed in seed in case of a Non-interactive PoRep is a fixed one. It was choosen
to be all zeros as such a byte array is easy to generate.

That fixed seed is used for the proving (and the internal aggregation) as well as
the verification.

This commit also removes the restriction that a seed cannot be all zeros, which
it also totally could be by coincidence (it's randomly generated in case of the
Interactive PoRep).