chore: move CI to GitHub Actions
Move the CI from CircleCI to GitHub actions.

As GitHub Actions work a bit differently, things needed to change a bit
and I took the chance to clean things up a bit. Nonetheless the coverage
should be the same.

If you compare it to the CircleCI configuration and wonder how different
things are, this list of changes might help:

 - Seperate multicore-sdr run was removed, it's enabled on the normal tests
 - Building benchmarks was removed as `--all-targets` already implies
   building the benchmarks
 - All the `test` jobs are removed, they are already covered by the release
   mode tests that run for the whole workspace
 - Some of the `ignored` tests were run separately. Now simply the whole
   workspace is run in a single job
 - The `no_gpu` test was renamed to `no-default-features`
 - `test_darwin` was renamed to `test_macos`
 - The `test_arm_no_gpu` wasn't ported as the MacOS runner is an M1, hence
   there is already basic ARM sanity checking